Switch to NATS as message broker
