+++ cmd/main.go
@@ -4,12 +4,15 @@ import (
 	"log"
 
 	"github.com/ilbarlo/flavourGeneratorConsumer/pkg/flavourmanager"
+	"github.com/nats-io/nats.go"
 )
+
+const natsURL = nats.DefaultURL
 
 func main() {
 	// Remember: configuration to run into a cluster
 
-	err := flavourmanager.StartConsumer("metrics", "amqp://guest:guest@localhost:5672/")
+	err := flavourmanager.StartConsumer("flavours", natsURL)
 	if err != nil {
 		log.Fatalf("failed to start consumer: %v", err)
 	}
