+++ go.mod
@@ -2,6 +2,14 @@ module github.com/ilbarlo/flavourGeneratorConsumer
 
 go 1.20
 
+require github.com/nats-io/nats.go v1.25.0
+
 require (
-	github.com/streadway/amqp v1.0.0
+	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/nats-io/nats-server/v2 v2.9.16 // indirect
+	github.com/nats-io/nkeys v0.4.4 // indirect
+	github.com/nats-io/nuid v1.0.1 // indirect
+	golang.org/x/crypto v0.8.0 // indirect
+	golang.org/x/sys v0.7.0 // indirect
+	google.golang.org/protobuf v1.30.0 // indirect
 )
